requirements.txt
openai==0.28.1
langchain==0.0.324
faiss-cpu==1.8.0
streamlit==1.28.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
setuptools==70.0.0
python-dotenv==1.0.1